requirements.txt
Flask
langchain
SQLAlchemy
python-dotenv 
langchain_community
google-generativeai
dateparser